Adding Video implementation
Right now I'm working on the ability for players to be use the milk filter with videos. The videos will be cut up frame by frame, have the filter applied, and then sown back together. So far I have the basic logic of the code ,but have not implemented the effects yet. There are other features I have in mind for the filter, but for now this is a good starting goal.

See ya next update!